Add simple implicit constructor invocation.
  Constructors are unique, receive no arguments and return no value,
and have the name of the base class, example:

	class	foo {
		foo() { ... }
	};

  When foo is instantiated, foo.foo() is called. If the class object
has initializers, it is done before calling the constructor, so,
the initialization can be overriden by the constructor.
  Superclass constructors are called implicitly. This was not done
in the previous language (unfinished) implementation, but all uses
of that required explicitly calling the superclass constructor, but,
since the derived class has the last word, this new approach is a
lot better, and does what a programmer using an object oriented language
would expect.