Add a single class/vector initializer token and syntax.
  The previous partial code used c99 like initializers, but the original
implementation used label like initializer, and this change reverts to
label like initializers (that are shorter).
  The c99 like initializers (.field = value, and [number] = value) may
lead to unreliable behavior due to dynamic initializers; but the same
is true if attempting to initialize one field with the value of another,
so, for now the error caused by attempting to use a variable before
full initialization is left in place, that is, one cannot write:
	a = { somefield: a.otherfield, otherfield: value };
because there are two initialization steps, one for static data and
another for dynamic data.