Properly make inplace hash operations work with base type keys.
  Initial code was only tested with "complex" key type, that did always
work due to values synchronized to the vm registers, and now properly
work with key values that are deferred until explicitly sync'ed to
memory, as they are small integers or floats. Also corrected invalid
sync when the operand should be changed to a hashentry after calling,
at runtime, ovm_gethash.