Better/proper constructor call implementation.
  The previous change was done in mistake, as it had already been
settled that it would use a least surprise mode implementation,
but the original code was checking the incorrect field of
orecord_t for the superclass (was checking the parent namespace).
  This means that constructors now cannot be explicitly called,
so, if there is some need for "communication" and initialization
order, it should be both, documented in the source and use
virtual methods to do so, also, one does not need to know if
a superclass implements a constructor or not before calling it
(neither can call it multiple times...).