Export almost every possible extra SDL2_mixer interface.
  This may still be redesigned a bit for better type checking, e.g. to
help users from causing errors due to assigning incorrect values to
channels.
  A new custom event type was also added, to generate an event when
a channel has finished playing. The new event returns the channel
identifier in the new channel field of event_t, and as the music one,
is a helper to avoid needing to poll to know when playing has finished.