Add several new surface_t utility builtin functions.
  These new builtins allow creating rgb and rgba surfaces (only accept
8 bits per channel), pulling a surface (that may be in some framebuffer
specific format) and pushing it back, as well as simple blitting helpers.
  New fields were added to the builtin surface_t object, to know the
masks (may be useful to write code without needing to know endianess),
bytes per pixel, and the pixels field that is only valid for the new
sdl_push_surface call and can be fetched with the new sdl_pull_surface
call.