Suspend thread in signal handler when not using semaphores
  This change is basically to use the same semantics of when
using semaphores. The previous implementation using sigsuspend
was not enabled by default due to race conditions, likely due
to asynchronously delivery of signals, that sometimes could be
sent to the wrong thread. Now it just blocks in the signal handler
and spins waiting for state change in a global volatile variable,
calling sched_yield until the variable changes value.
  The new behavior has been made default.